159 - Rename folder and files and separate SeriesList from SeriesCard
