+++ frontend/src/components/card/SeriesCard.styles.js
@@ -1,34 +1,5 @@
 ﻿import styled from 'styled-components';
 
-export const SelectedTeamContainer = styled.div`
-    background: var(--hermanni-grey-lighter);
-
-    .selectedTeamHeading {
-        flex: 1 100%;
-        flex-direction: row;
-        display: flex;
-        align-items: center;
-        justify-content: space-between;
-
-        span {
-            font-size: 12px;
-            color: var(--evidence-grey);
-            vertical-align: -webkit-baseline-middle;
-        }
-    }
-`;
-
-export const CardContainer = styled.div`
-    display: grid;
-    grid-gap: 16px;
-    grid-template-columns: repeat(
-        auto-fit,
-        minmax(calc(var(--max-page-width) / 4 - var(--space-16)), 1fr)
-    );
-    width: 100%;
-    height: 100%;
-`;
-
 export const CardSection = styled.section`
     background-color: var(--nero-white);
     box-shadow: 0px 2px 4px var(--hermanni-grey);
@@ -106,8 +77,4 @@ export const CardValue = styled.span`
 export const StatusSpan = styled.span`
     color: ${props =>
         props.status === 'FAIL' ? 'var(--nelson-purple)' : 'var(--pirlo-blue)'};
-`;
-
-export const SeriesCount = styled.span`
-    margin-right: 16px;
 `;
