+++ frontend/src/components/card/TeamCard.styles.js
