+++ frontend/src/pages/SeriesList.jsx
@@ -1,0 +1,76 @@
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import { useLocation } from 'react-router-dom';
+import { useStateValue } from '../contexts/state';
+import BreadcrumbNav from '../components/BreadcrumbNav';
+import { ContainerGrid12, ContentGrid6 } from '../styles/baseComponents';
+import {
+    CardContainer,
+    SelectedTeamContainer,
+    SeriesCount,
+} from './SeriesList.styles';
+import NotFound from '../components/NotFound';
+import PropTypes from 'prop-types';
+import SeriesCard from '../components/card/SeriesCard';
+
+const SeriesList = ({ selectedTeam }) => {
+    const [t] = useTranslation(['team']);
+    let location = useLocation();
+    const teamName = location.pathname.substring(
+        location.pathname.lastIndexOf('/') + 1
+    );
+
+    const [{ teamsState }] = useStateValue();
+    const seriesCount = teamsState.find(team => team.name.includes(teamName))
+        .series_count;
+
+    return (
+        <div id="selectedTeam">
+            <BreadcrumbNav status={'team'} />
+            {selectedTeam && selectedTeam.all_builds ? (
+                <>
+                    <ContainerGrid12>
+                        <ContentGrid6>
+                            <h1>Team {teamName}</h1>
+                        </ContentGrid6>
+                    </ContainerGrid12>
+                    <SelectedTeamContainer>
+                        <ContainerGrid12>
+                            <div className={'selectedTeamHeading'}>
+                                <h2>
+                                    {t('card.last_build.header')} {teamName}
+                                </h2>
+                                <SeriesCount>
+                                    {seriesCount} {t('card.last_build.series')}{' '}
+                                </SeriesCount>
+                            </div>
+                            <CardContainer>
+                                <SeriesCard data={selectedTeam.all_builds} />
+                                {selectedTeam.series
+                                    .reverse()
+                                    .map((serie, i) => {
+                                        return (
+                                            <SeriesCard key={i} data={serie} />
+                                        );
+                                    })}
+                            </CardContainer>
+                        </ContainerGrid12>
+                    </SelectedTeamContainer>
+                </>
+            ) : (
+                <NotFound />
+            )}
+        </div>
+    );
+};
+
+SeriesList.propTypes = {
+    selectedTeam: PropTypes.shape({
+        all_builds: PropTypes.object,
+        name: PropTypes.string,
+        series: PropTypes.array,
+        series_count: PropTypes.number,
+    }),
+};
+
+export default SeriesList;
